+++ summingbird-builder/src/main/scala/com/twitter/summingbird/builder/CompletedBuilder.scala
@@ -36,10 +36,10 @@ import java.io.Serializable
 
 object CompletedBuilder {
   def injectionRegistrar[T: Manifest](injection: Injection[T, Array[Byte]]) =
-    InjectionRegistrar(manifest[T].erasure.asInstanceOf[Class[T]], injection)
+    InjectionRegistrar(manifest[T].runtimeClass.asInstanceOf[Class[T]], injection)
 
   def injectionDefaultRegistrar[T: Manifest](injection: Injection[T, Array[Byte]]) =
-    InjectionDefaultRegistrar(manifest[T].erasure.asInstanceOf[Class[T]], injection)
+    InjectionDefaultRegistrar(manifest[T].runtimeClass.asInstanceOf[Class[T]], injection)
 }
 
 case class CompletedBuilder[P <: Platform[P], K, V](