+++ summingbird-online/src/main/scala/com/twitter/summingbird/online/executor/FinalFlatMap.scala
@@ -16,8 +16,7 @@ limitations under the License.
 
 package com.twitter.summingbird.online.executor
 
-import com.twitter.algebird.{ SummingQueue, Semigroup, MapAlgebra }
-import com.twitter.algebird.util.summer.AsyncSummer
+import com.twitter.algebird.Semigroup
 import com.twitter.bijection.Injection
 import com.twitter.util.Future
 
@@ -25,15 +24,12 @@ import com.twitter.summingbird.online.Externalizer
 
 import com.twitter.summingbird.online.FlatMapOperation
 
-import com.twitter.summingbird.option.CacheSize
 import com.twitter.summingbird.online.option.{
   SummerBuilder,
   MaxWaitingFutures,
   MaxFutureWaitTime,
-  MaxEmitPerExecute,
-  FlushFrequency
+  MaxEmitPerExecute
 }
-import scala.collection.breakOut
 import scala.collection.mutable.{ Map => MMap, ListBuffer }
 // These CMaps we generate in the FFM, we use it as an immutable wrapper around
 // a mutable map.